firebase/functions/requirements.txt
firebase_functions~=0.1.0
python-dotenv~=1.0.1
openai~=1.35.5
